Added overwrite control to copy run files

+++ gggutils/i2srun.py
@@ -178,7 +178,7 @@ def _make_new_i2s_run_file(datestr, run_files, last_key_with_file, save_dir, ove
     shutil.copy2(run_files[last_key_with_file], new_file)
 
 
-def copy_i2s_run_files_from_target_dirs(dirs_list, save_dir, interactive='choice'):
+def copy_i2s_run_files_from_target_dirs(dirs_list, save_dir, interactive='choice', overwrite=None):
     avail_target_dates = target_utils.build_target_dirs_dict([], dirs_list=dirs_list, flat=True,
                                                              full_datestr=True, key_by_basename=False)
     for site, site_dict in avail_target_dates.items():
@@ -215,8 +215,16 @@ def copy_i2s_run_files_from_target_dirs(dirs_list, save_dir, interactive='choice
             prefix = 'opus-i2s' if re.match('opus', os.path.basename(input_file)) else 'slice-i2s'
             new_base_filename = '{}.{}.in'.format(prefix, site_date)
             new_fullname = os.path.join(save_dir, new_base_filename)
-            logger.info('Copying {} to {}'.format(input_file, new_fullname))
-            shutil.copy(input_file, new_fullname)
+            if os.path.exists(new_fullname) and overwrite is None:
+                this_ow = uielements.user_input_yn('{} exists. Overwrite?'.format(new_fullname))
+            else:
+                this_ow = overwrite
+
+            if os.path.exists(new_fullname) and not this_ow:
+                logger.debug('Not copying {} to {} - exists'.format(input_file, new_fullname))
+            else:
+                logger.info('Copying {} to {}'.format(input_file, new_fullname))
+                shutil.copy(input_file, new_fullname)
 
 
 def link_i2s_input_files(cfg_file, overwrite=False, clean_links=False, clean_spectra=False):
@@ -868,6 +876,10 @@ def parse_copy_i2s_target_runfiles_args(parser):
                         help='Alter interactive behavior. "choice" will prompt the user to choose a file if multiple '
                              'files are found. "all" will always prompt, no matter how many are found, and "none" will '
                              'never prompt. Note that if multiple files are found, "none" will cause an error.')
+    parser.add_argument('-o', '--overwrite', action='store_true', default=None,
+                        help='Overwrite destination file if it exists. The default behavior is to ask.')
+    parser.add_argument('-n', '--no-overwrite', action='store_false', default=None,
+                        help='Never overwrite destination files if they exist. The default behavior is to ask.')
     parser.set_defaults(driver_fxn=copy_i2s_run_files_from_target_dirs)
